requirements.txt
Django==4.1.3
psycopg2-binary==2.9.9
python-decouple==3.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
